+++ src/hoc/PrivateRouter.jsx
@@ -1,16 +1,22 @@
 import { Navigate, useLocation } from 'react-router-dom';
-import { useSelector } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import PropTypes from 'prop-types';
 import { userSelector } from '../services/selectors/userSelector';
+
 
 
 const PrivateRouter = ({ children }) => {
 
     const location = useLocation()
 
-    const { token,auth } = useSelector(userSelector)  
+    const dispatch = useDispatch()
 
-    return token||auth ? children : <Navigate to='/login' state={{ from: location }} />
+    const { token,auth } = useSelector(userSelector)    
+
+  
+ 
+
+ return token||auth ? children : <Navigate to='/login' state={{ from: location }} replace />
 }
 
 PrivateRouter.propTypes={
