+++ src/hoc/PublicRouter.jsx
@@ -7,18 +7,17 @@ import { userSelector } from '../services/selectors/userSelector';
 const PublicRouter = ({ children }) => {
 
   const location = useLocation()
+    
+
+  
+
   const { token, auth } = useSelector(userSelector)
 
-
-  // const path = location?.pathname;
-
-  // if(path==='/fargot'||path==='/reset'){
-  //     return <Navigate to={'/login'}  replace/>
-  // }
-
+ 
+  
   return (
     <>
-      {!token ? (children) : <Navigate to='/home' state={{ from: location }} />}
+      {!token||!auth ? (children) : <Navigate to='/home' replace/>}
     </>
 
   )
