+++ src/pages/Login/login.scss
@@ -46,6 +46,7 @@
         color: #8585AD;
         display: inline-flex;
         gap: 10px;
+        align-items: center;
     }
 
     &__btn {
