+++ src/pages/ResetPassword/ResetPage.jsx
@@ -1,12 +1,11 @@
 import React, { useCallback, useState } from 'react'
 import { useDispatch } from 'react-redux'
-import {toast} from 'react-toastify'
+import { toast } from 'react-toastify'
 import { useNavigate } from 'react-router-dom'
 import { passSchema, restCodeSchema, checkValidate } from '../../components/validation'
 import { Input, Button } from '@ya.praktikum/react-developer-burger-ui-components'
-
+import { resetPassword } from '../../services/actions/user'
 import './reset.scss'
-import { resetPassword } from '../../services/actions'
 
 const ResetPage = () => {
 
@@ -38,10 +37,10 @@ const ResetPage = () => {
 
   const handleSubmit = useCallback((e) => {
     e.preventDefault()
-    dispatch(resetPassword(reset.password,toast,navigate))
+    dispatch(resetPassword(reset.password, toast, navigate))
     setReset({ ...reset, password: '', code: '' })
 
-  }, [reset,navigate])
+  }, [reset, navigate,dispatch])
 
   return (
     <div className='reset'>
@@ -86,7 +85,7 @@ const ResetPage = () => {
         </Button>
         <div className='fargot__footer'>
           <p className='fargot__text'>Вспомнили пароль?
-            <button className='fargot__btn'  onClick={backToLogin}>Войти</button>
+            <button className='fargot__btn' onClick={backToLogin}>Войти</button>
           </p>
         </div>
       </form>
