+++ src/services/actions/user.js
@@ -1,44 +1,14 @@
+
 import {
     API_GET_USER,
     API_USER_LOGIN, API_USER_LOGOUT,
     API_USER_RECOVER, API_USER_REGISTER,
     API_USER_RESET,
     API_USER_TOKEN,
-    INGREDIENTS_URL,
-    ORDERS_URL
-} from "../../constants"
+   } from '../../constants'
+
 import { deleteAllCookies, getCookie, setCookie } from "../../utils/cookies"
 import { checkResponse, getData } from "../../utils/data"
-
-
-// Получение ингридиентов
-export const GET_INGREDIENTS_SUCCES = 'GET_INGREDIENTS_SUCCES'
-export const GET_INGREDIENTS_REQUEST = 'GET_INGREDIENTS_REQUEST'
-export const GET_INGREDIENTS_ERROR = 'GET_INGREDIENTS_ERROR'
-
-// Получение заказа
-export const POST_ORDER_REQUEST = 'GET_ORDER_REQUEST'
-export const POST_ORDER_SUCCESS = 'GET_ORDER_SUCCESS'
-export const POST_ORDER_ERROR = 'GET_ORDER_ERROR'
-export const SET_LOGIN_CHECKED = 'SET_LOGIN_CHECKED'
-export const CLEAR_ORDER_NUMBER = 'CLEAR_ORDER_NUMBER,'
-
-// Ингредиенты конструктора
-
-export const GET_INGREDIENTS_CONSTRUCTOR = 'GET_INGREDIENTS_CONSTRUCTOR';
-export const POST_BUN_CONSTRUCTOR = 'POST_BUN_CONSTRUCTOR';
-export const POST_FILLING_CONSTRUCTOR = 'POST_ELEMENT_CONSTRUCTOR';
-export const DELETE_FILLING_CONSTRUCTOR = 'DELETE_ELEMENT_CONSTRUCTOR';
-export const UPDATE_LIST_FILLINGS = 'UPDATE_LIST_FILLINGS';
-
-// Открытие и закрытие модалки
-
-export const OPEN_MODAL_INGREDIENT = 'OPEN_MODAL_INGREDIENT';
-export const CLOSE_MODAL = 'CLOSE_MODAL_INGREDIENT';
-export const OPEN_MODAL_ORDER = 'OPEN_MODAL_ORDER';
-export const CLEAR_ORDER_LIST = 'CLEAR_ORDER_LIST';
-
-//Post Users
 
 export const POST_USER_REQUEST = 'POST_USER_REQUEST';
 export const POST_USER_SUCCES = 'POST_USER_SUCCESS';
@@ -59,8 +29,6 @@ export const UPDATE_USER_INFO = 'UPDATE_USER_INFO';
 export const UPDATE_USER_INFO_REQUEST = 'UPDATE_USER_INFO_REQUEST';
 export const UPDATE_USER_INFO_FAILED = 'UPDATE_USER_INFO_FAILED';
 
-//passwordS
-
 export const FARGOT_PASSWORD_SUCCESFLY = 'FARGOT_PASSWORD_SUCCESFLY';
 
 export const FARGOT_PASSWORD_REQUEST = 'FARGOT_PASSWORD_REQUEST';
@@ -69,97 +37,36 @@ export const FARGOT_PASSWORD_FAILED = 'FARGOT_PASSWORD_FAILED';
 
 export const SET_FARGOT_CHECKED = 'SET_FARGOT_CHECKED';
 
+export const SET_LOGIN_CHECKED = 'SET_LOGIN_CHECKED';
+
 
 export const RESET__PASSWORD_SUCCESSFLY = 'RESET_PASSWORD_SUCCESSFLY';
 
 export const RESET__PASSWORD_FAILED = 'RESET_PASSWORD_FAILED';
 
 
-
-
-// Actions Get All Ingridients
-
-export const getIngredients = () => async dispatch => {
-
-    dispatch({ type: GET_INGREDIENTS_REQUEST })
-    try {
-        const { data } = await getData(INGREDIENTS_URL)
-        dispatch({ type: GET_INGREDIENTS_SUCCES, payload: data })
-
-    } catch (error) {
-        console.log(error.message)
-        dispatch({ type: GET_INGREDIENTS_ERROR, error: error.message })
-
-    }
-
-}
-
-//Actions Get Order Number
-
-export const getOrderNumber = (burgersIds) => async dispatch => {
-
-    dispatch({ type: POST_ORDER_REQUEST })
-    try {
-
-    const { order, success } = await getData(ORDERS_URL, 'POST', { ingredients: [...burgersIds] })
-        if (success) {
-            dispatch({ type: POST_ORDER_SUCCESS, payload: order })
-
-
-        }
-
-    } catch (error) {
-        console.log(error.message)
-        dispatch({ type: POST_ORDER_ERROR, payload: error.message })
-
-    }
-
-}
-
-//actions Constructor
-
-export const getElementsByConstructor = () => ({
-    type: GET_INGREDIENTS_CONSTRUCTOR
-})
-
-export const addBunByConstructor = (bun) => ({
-    type: POST_BUN_CONSTRUCTOR,
-    payload: bun
-})
-
-export const addFillingConstructor = (item) => ({
-    type: POST_FILLING_CONSTRUCTOR,
-    payload: item
-})
-
-export const deleteFillingOfConstructor = (itemId) => ({
-    type: DELETE_FILLING_CONSTRUCTOR,
-    payload: itemId
-})
-
-
-//GET Ingridient of Modal
-export const openModalIngredient = (element) => ({
-    type: OPEN_MODAL_INGREDIENT,
-    payload: element
-})
-
-
-export const closeModal = () => dispatch => {
-    dispatch({ type: CLOSE_MODAL })
-    dispatch({ type: CLEAR_ORDER_LIST })
-}
-
-////////////////////Users/////////////////////////////////////////////////////
-
-//register
-
+/**
+ * Функция регистрации
+ */
 export const registerUser = (body, navigate, toast) => async dispatch => {
 
-    dispatch({ type: POST_USER_REQUEST, payload: body })
-    try {
+     console.log(body)
 
-        const data = await getData(API_USER_REGISTER, 'POST', { ...body })
+    dispatch({ type: POST_USER_REQUEST, payload: body })
+
+    try {    
+
+     const res = await fetch(API_USER_REGISTER, {
+            method: 'POST',
+            headers: {
+                "Content-Type": "application/json"
+            },
+            body: JSON.stringify(body)
+        })
+
+        checkResponse(res)
+
+     const data = await res.json()
 
         if (data.success) {
             toast.success('Регистрация прошла успешно!!')
@@ -176,10 +83,18 @@ export const registerUser = (body, navigate, toast) => async dispatch => {
     }
 }
 
-//login
+/**
+ * Функция авторизации
+ * @param {Objeсt} body 
+ * @param {Hook} navigate 
+ * @param {Notification} toast 
+ * @param {String Path} fromPage 
+ * @returns Object data.user
+ */
 
 export const loginUser = (body, navigate, toast, fromPage) => async dispatch => {
 
+    
     dispatch({ type: LOGIN_USER_REQUEST })
 
     try {
@@ -196,43 +111,66 @@ export const loginUser = (body, navigate, toast, fromPage) => async dispatch => 
 
         if (!res.ok) {
             dispatch({ type: SET_FARGOT_CHECKED })
-            throw new Error('Что-то пошло не так...')
+            throw new Error('Проверьте логин или пароль')
+
         }
 
         const data = await res.json()
 
 
+
         if (data.success) {
             toast.success('Вы вошли в систему!!')
             dispatch({ type: LOGIN_USER_SUCCESS, payload: data.user })
-            dispatch({type:SET_LOGIN_CHECKED})
+            dispatch({ type: SET_LOGIN_CHECKED })
             setCookie('accessToken', data.accessToken.split(' ')[1]);
-            setCookie('refreshToken', data.refreshToken);
+            setCookie('refreshToken', data.refreshToken)         
             navigate(`${fromPage}`)
+
         }
+
 
     } catch (error) {
         dispatch({ type: POST_USER_FAILED, payload: error.message })
-        toast.error('Что-то пошло не так...')
+        toast.error(error.message)
 
     }
 }
 
+/**
+ * Функция выхода
+ * @param {Notifaction} toast 
+ * @param {Hook} navigate 
+ * @returns 
+ */
 
-//logout
 export const logout = (toast, navigate) => async dispatch => {
 
     try {
-        const data = await getData(API_USER_LOGOUT, 'POST', {
-            token: getCookie('refreshToken')
+
+        const res = await fetch(API_USER_LOGOUT, {
+            method: 'POST',
+            headers: {
+                "Content-Type": "application/json"
+            },
+            body: JSON.stringify({ token: getCookie('refreshToken') })
         })
+
+        checkResponse(res)
+
+        const data = await res.json()
+
+        console.log(data)
+
         if (data.success) {
             toast.success('Вы вышли из из системы!')
             dispatch({ type: LOGOUT_USER })
             deleteAllCookies()
+            window.localStorage.removeItem('user')
             navigate('/login')
 
         }
+
 
     } catch (error) {
         console.log(error.message)
@@ -240,8 +178,10 @@ export const logout = (toast, navigate) => async dispatch => {
 
 }
 
-//updateToken
-
+/**
+ * Функция обновления токена
+ * @returns token
+ */
 const updateToken = () => async dispatch => {
 
     try {
@@ -253,6 +193,8 @@ const updateToken = () => async dispatch => {
         if (data.success) {
             setCookie('accessToken', data.accessToken.split(' ')[1]);
             setCookie('refreshToken', data.refreshToken);
+            dispatch(getUser())
+
         }
 
     } catch (error) {
@@ -262,9 +204,10 @@ const updateToken = () => async dispatch => {
 
 }
 
-
-//getUser
-
+/**
+ * Функция получения пользователя из БД
+ * @returns 
+ */
 export const getUser = () => async dispatch => {
 
     dispatch({ type: GET_USER_REQUEST })
@@ -278,19 +221,21 @@ export const getUser = () => async dispatch => {
                     : ''
             }
         })
-        if (!res.status.ok) {
-            dispatch(updateToken())
-        }
-        checkResponse(res)
-        const data = await res.json()
 
+        checkResponse(res)
+
+        const data = await res.json()
 
         if (data.success) {
             dispatch({ type: GET_USER_INFO, payload: data.user })
 
+
         }
 
     } catch (error) {
+        if (error.message === 'jwt expired' || 'jwt malformed') {
+            dispatch(updateToken())
+        }
         console.log(error.message)
         dispatch({ type: GET_USER_FAILED, payload: error.message })
     }
@@ -299,7 +244,12 @@ export const getUser = () => async dispatch => {
 
 
 
-//update user info
+/**
+ * Функция обновления данных 
+ * @param {Object} form Данные от клиента
+ * @param {Notificaton} toast 
+ * @returns data.user Новые данные 
+ */
 
 export const updateUserInfo = (form, toast) => async dispatch => {
 
@@ -320,9 +270,12 @@ export const updateUserInfo = (form, toast) => async dispatch => {
 
         const data = await res.json()
 
-        if (data.success) {
+
+
+        if (data?.success) {
             toast.success('Данные успешно обновлены!!')
             dispatch({ type: UPDATE_USER_INFO, payload: data.user })
+
         }
 
     } catch (error) {
@@ -333,7 +286,15 @@ export const updateUserInfo = (form, toast) => async dispatch => {
 
 }
 
-//Fargot Password
+
+/**
+ * Функция сброса пароля
+ * @param {String} email 
+ * @param {Hook} navigate 
+ * @param {Notification} toast 
+ * @returns 
+ */
+
 export const fargotPassword = (email, navigate, toast) => async dispatch => {
 
     dispatch({ type: FARGOT_PASSWORD_REQUEST })
@@ -341,12 +302,12 @@ export const fargotPassword = (email, navigate, toast) => async dispatch => {
     try {
 
         const data = await getData(API_USER_RECOVER, 'POST', { email })
-        console.log(data)
+
 
         if (data.success) {
             toast.success('Это успех!!')
             dispatch({ type: FARGOT_PASSWORD_SUCCESFLY, payload: data?.message })
-            navigate('/reset')
+            navigate('/reset-password')
         }
     } catch (error) {
         console.log(error.message)
@@ -356,7 +317,14 @@ export const fargotPassword = (email, navigate, toast) => async dispatch => {
 
 }
 
-//Reset Password
+/**
+ * Функия обновления пароля
+ * @param {string} password 
+ * @param {notification} toast 
+ * @param {hook} navigate 
+ * @returns 
+ */
+
 export const resetPassword = (password, toast, navigate) => async dispatch => {
 
     try {
@@ -377,7 +345,3 @@ export const resetPassword = (password, toast, navigate) => async dispatch => {
     }
 
 }
-
-
-
-
